requirements.txt
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-restx==1.0.3
greenlet==2.0.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
pkgutil-resolve-name==1.3.10
PyJWT==2.6.0
pyrsistent==0.19.1
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.42
Werkzeug==2.2.2
zipp==3.10.0
